+++ src/infrastructure/http/controllers/index.js
@@ -1,0 +1,11 @@
+/**
+ * HTTP Controllers
+ *
+ * Controllers that map HTTP requests to application use cases.
+ *
+ * @module infrastructure/http/controllers
+ */
+
+export { StationController } from './StationController.js';
+export { PlatformController } from './PlatformController.js';
+export { InstrumentController } from './InstrumentController.js';